move test test to utility module
