+++ utility/build.gradle.kts
@@ -10,9 +10,12 @@ repositories {
 }
 
 dependencies {
+    implementation(project(":core"))
+    implementation("org.apache.commons:commons-math3:3.6.1")
+
     testImplementation(platform("org.junit:junit-bom:5.10.0"))
     testImplementation("org.junit.jupiter:junit-jupiter")
-    implementation("org.apache.commons:commons-math3:3.6.1")
+    testImplementation(kotlin("test"))
 }
 
 tasks.test {
